requirements.txt
ipywidgets
matplotlib
numpy
pandas
tqdm
python-dotenv
openai
tiktoken